feat: DEV-390 Refactor more
