+++ src/lib/calculateValueByPercent.js
@@ -1,0 +1,5 @@
+export const calculateValueByPercent = (percentage, value, toFixed = 2) => {
+  const calculation = Math.round(((percentage / 100) * value).toFixed(toFixed))
+
+  return calculation
+}
