+++ src/lib/convertToMoney.js
@@ -1,0 +1,3 @@
+export const convertToMoney = (number) => {
+  return number.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',')
+}
